[task] update cmt

+++ ampro015-service/src/main/java/com/poscodx/odc/ampro015/service/rest/Pme00TaskResource.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.poscdx.odc.ampro015.domain.entity.M00EmployeeTaskId;
 import com.poscdx.odc.ampro015.domain.entity.M00Task;
+import com.poscdx.odc.ampro015.domain.entity.M00TaskDto;
 import com.poscdx.odc.ampro015.domain.entity.M00TaskId;
 import com.poscdx.odc.ampro015.domain.entity.Pme00EmployeeTask;
 import com.poscdx.odc.ampro015.domain.lifecycle.ServiceLifecycle;
@@ -13,6 +14,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -34,8 +36,15 @@ public class Pme00TaskResource {
 
     @CrossOrigin
     @GetMapping(path = "/getAll")
-    public List<M00Task> findAll(@RequestParam String projectNumber, @RequestParam(defaultValue = "0", required = false) int pageNo, @RequestParam(defaultValue = "20", required = false) int pageSize, @RequestParam(defaultValue = "TASK_NAME", required = false) String sortBy, @RequestParam(defaultValue = "ASC", required = false) String sortDirection) {
-        return this.serviceLifecycle.requestTaskService().findAll(projectNumber, pageNo, pageSize, sortBy, sortDirection);
+    public List<M00TaskDto> findAll(@RequestParam String projectNumber,
+                                    @RequestParam(required = false, defaultValue = "") String taskName,
+                                    @RequestParam(required = false, defaultValue = "") String planDate,
+                                    @RequestParam(required = false, defaultValue = "") String actualEndDate,
+                                    @RequestParam(defaultValue = "0", required = false) int pageNo,
+                                    @RequestParam(defaultValue = "20", required = false) int pageSize,
+                                    @RequestParam(defaultValue = "TASK_NAME", required = false) String sortBy,
+                                    @RequestParam(defaultValue = "ASC", required = false) String sortDirection) {
+        return this.serviceLifecycle.requestTaskService().findAll(projectNumber, taskName, planDate, actualEndDate, pageNo, pageSize, sortBy, sortDirection);
     }
 
     @CrossOrigin
@@ -50,24 +59,15 @@ public class Pme00TaskResource {
     public Map insertTask(@RequestBody Map<String, Object> newTaskRequest) {
 
         //convert json to DTO
-        M00Task newTask = objectMapper.convertValue(newTaskRequest.get(TASK_KEY), M00Task.class);
-        List<Pme00EmployeeTask> pme00EmployeeTask = new ArrayList<>();
-
-        List<Object> pme00EmployeeTasksList = objectMapper.convertValue(newTaskRequest.get(MEMBERS_KEY), List.class);
-
-        pme00EmployeeTasksList.forEach(pme00EmployeeTask1 -> {
-            pme00EmployeeTask.add(objectMapper.convertValue(pme00EmployeeTask1, Pme00EmployeeTask.class));
-        });
         //insert task
-        M00Task newM00TaskRegisted = this.serviceLifecycle.requestTaskService().register(newTask);
+
 
         //inset to task member
-        List<Pme00EmployeeTask> pme00EmployeeTaskResponse = this.serviceLifecycle.requestPme00EmployeeTaskService().createFromList(pme00EmployeeTask);
 
         // return new list
         Map response = new HashMap<>();
-        response.put(TASK_KEY, newM00TaskRegisted);
-        response.put(MEMBERS_KEY, pme00EmployeeTaskResponse);
+        response.put(TASK_KEY, "");
+        response.put(MEMBERS_KEY, "");
 
         return response;
     }
@@ -122,15 +122,24 @@ public class Pme00TaskResource {
         Optional<M00Task> existedTask = Optional.ofNullable(this.serviceLifecycle.requestTaskService().find(m00TaskId));
         if (existedTask.isPresent()) {
             //delete data tb_pme00_employee_task
-            M00EmployeeTaskId requestM00EmployeeTaskId = new M00EmployeeTaskId()
-            this.serviceLifecycle.requestPme00EmployeeTaskService().remove();
-            this.serviceLifecycle.requestTaskService().remove(m00TaskId);
-            Map response = new HashMap<>();
-            response.put(DATA_KEY, "");
-            response.put(DESCRIPTION_KEY, "Delete task success!!!");
-            response.put(STATUS_KEY, HttpStatus.BAD_REQUEST.value());
-            return ResponseEntity.ok(response);
-        } else {
+            List<Pme00EmployeeTask> pme00EmployeeTasksExisted = this.serviceLifecycle.requestPme00EmployeeTaskService().findAllByTaskId(m00TaskId);
+            if (!pme00EmployeeTasksExisted.isEmpty()) {
+                //M00EmployeeTaskId requestM00EmployeeTaskId = new M00EmployeeTaskId(existedTask.get().getProjectNumber(), existedTask.get().getTaskName(), existedTask.get().getEmpId());
+                this.serviceLifecycle.requestPme00EmployeeTaskService().removeByListEmployeeTask(pme00EmployeeTasksExisted);
+                this.serviceLifecycle.requestTaskService().remove(m00TaskId);
+                Map response = new HashMap<>();
+                response.put(DATA_KEY, "");
+                response.put(DESCRIPTION_KEY, "Delete task success!!!");
+                response.put(STATUS_KEY, HttpStatus.BAD_REQUEST.value());
+                return ResponseEntity.ok(response);
+            } else {
+                Map response = new HashMap<>();
+                response.put(DATA_KEY, "");
+                response.put(DESCRIPTION_KEY, "Can't delete task success!!!");
+                response.put(STATUS_KEY, HttpStatus.BAD_REQUEST.value());
+                return ResponseEntity.badRequest().body(response);
+            }
+        } else{
             Map response = new HashMap<>();
             response.put(DATA_KEY, "");
             response.put(DESCRIPTION_KEY, "Can't delete task success!!!");
@@ -138,5 +147,4 @@ public class Pme00TaskResource {
             return ResponseEntity.badRequest().body(response);
         }
     }
-
 }
